Use a int64_t for nMaxStakeValue.

+++ src/init.cpp
@@ -43,7 +43,7 @@ bool fMinimizeCoinAge;
 unsigned int nNodeLifespan;
 unsigned int nDerivationMethodIndex;
 unsigned int nMinerSleep;
-unsigned int nMaxStakeValue;
+int64_t nMaxStakeValue;
 bool fUseFastIndex;
 enum Checkpoints::CPMode CheckpointsMode;
